pyproject.toml
[project]
name = "kb-cli"
version = "0.1.0"
description = "Personal knowledge base CLI with Telegram channel publishing"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "trafilatura>=1.6.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pypdf>=4.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-genai>=1.0.0",
]

[project.scripts]
kb = "kb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kb"]
